OD-11593 Regenerate code from adjusted spec file. Change numeric type to integer. Remove useless api: /getPageRenderer
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@50348 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ editor-api/src/main/groovy/ish/oncourse/willow/editor/model/Block.java
@@ -3,7 +3,7 @@ package ish.oncourse.willow.editor.model;
 
 public class Block  {
   
-    private Double id = null;
+    private Integer id = null;
     private String title = null;
     private String content = null;
 
@@ -11,21 +11,21 @@ public class Block  {
      * Get id
      * @return id
      */
-    public Double getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Double id) {
+    public void setId(Integer id) {
        this.id = id;
     }
 
-    public Block id(Double id) {
+    public Block id(Integer id) {
       this.id = id;
       return this;
     }
 
     /**
-     * Title/Name of block
+     * Unique name of block
      * @return title
      */
     public String getTitle() {
@@ -42,7 +42,7 @@ public class Block  {
     }
 
     /**
-     * Html source of block
+     * Rich text content
      * @return content
      */
     public String getContent() {
