+++ editor-api/src/main/groovy/ish/oncourse/willow/editor/rest/WebNodeToPage.groovy
@@ -20,8 +20,8 @@ class WebNodeToPage {
         return new Page().with {page ->
             page.title = webNode.name
             page.reservedURL = false
-            page.number = webNode.nodeNumber.doubleValue()
-            page.themeId = webNode.webNodeType.id.doubleValue()
+            page.number = webNode.nodeNumber
+            page.themeId = webNode.webNodeType.id.intValue()
             page.content = webNode.webContentVisibility?.find {it.regionKey == RegionKey.content}?.webContent?.contentTextile
             page.visible = webNode.published
             page.urls += webNode.webUrlAliases .collect { webUrlAlias ->  WebUrlAliasToPageUrl.valueOf(webUrlAlias).pageUrl }
